Do some cleanup and move things around
