+++ ndcube/extra_coords/__init__.py
@@ -1,3 +1,4 @@
 from .extra_coords import ExtraCoords
+from .table_coord import *
 
-__all__ = ['ExtraCoords']
+__all__ = ['ExtraCoords', 'TimeTableCoordinate', 'SkyCoordTableCoordinate', 'QuantityTableCoordinate']
