--- ndcube/global_coords/__init__.py
@@ -1,3 +1,0 @@
-from .global_coords import GlobalCoords
-
-__all__ = ['GlobalCoords']
